@@ -1,7 +1,6 @@
 from __future__ import absolute_import
 # Copyright (c) 2010-2019 openpyxl
 
-from openpyxl.compat import unicode
 from openpyxl.descriptors.serialisable import Serialisable
 from openpyxl.descriptors import (
     Alias,
@@ -20,7 +19,7 @@ class PivotSource(Serialisable):
 
     tagname = "pivotSource"
 
-    name = NestedText(expected_type=unicode)
+    name = NestedText(expected_type=str)
     fmtId = NestedInteger(expected_type=int)
     extLst = Typed(expected_type=ExtensionList, allow_none=True)
 
