@@ -1,9 +1,6 @@
-from __future__ import absolute_import
 # Copyright (c) 2010-2019 openpyxl
 
 import pytest
-
-from openpyxl.compat import unicode
 
 
 @pytest.fixture
@@ -88,4 +85,4 @@ class TestReference:
 
     def test_repr(self, Reference):
         ref = Reference(range_string=b"'D\xc3\xbcsseldorf'!A1:A10".decode("utf8"))
-        assert unicode(ref) == b"'D\xc3\xbcsseldorf'!$A$1:$A$10".decode("utf8")
+        assert str(ref) == b"'D\xc3\xbcsseldorf'!$A$1:$A$10".decode("utf8")
